Add apparmor enabled flag (#37)
This adds a flag that allows users to control whether apparmor
enforcement is configured. Some environments (such as Amazon Linux 2)
do not support apparmor enforcement.

Removal of the apparmor annotations on the PSP is required because if
they are included, pods will start up with a state of `Blocked` because
of the lack of ability to enforce apparmor in these environments.
Removal of the apparmor annotations on the DaemonSet is required because
otherwise the pods that are launched don't have a PSP they can select.